Add warning & redirect txIds to dispute details & filter
Make sure the published warning & redirect txIds of a refund dispute
show up in the details window, in place of the DPT ID for v4 protocol
trades. (Unlike the latter, the fields are missing from the details for
mediation disputes, as it isn't known which combination of staged txs
will be published, if any.)

Also make sure the two txIds are filterable on, by adding them to the
'DisputeView.FilterResult' enum and 'getFilterResult(..)' method.