Restructure redeem script to reduce size of redirect & claim txs
Deduplicate the public key used in the claim path of the warning escrow
output redeem script, of the v5 protocol, reducing the size of the
redirect & claim txs by 8 vbytes (32 WUs) each. The claim public key
doubles up as the trader's 2-of-2 multisig public key, used for both the
regular escrow output of the deposit tx and the redirection spend path
of the warning escrow output, so it may be reused through some
manipulation of the bitcoin stack.

Also reverse the order of the buyer & seller pubKeys & signatures, so
that the buyer signature is always higher up in the stack, for
consistency with the deposit tx redeem script.

This leads to the following two warning escrow output redeem scripts,
structured slightly differently:

Buyer's warning escrow output redeem script:

  <buyerPubKey> OP_SWAP
  OP_IF
    2 <sellerPubKey> OP_ROT 2 OP_CHECKMULTISIG
  OP_ELSE
    <claimDelay> OP_CHECKSEQUENCEVERIFY OP_DROP OP_CHECKSIG
  OP_ENDIF

Seller's warning escrow output redeem script:

  <sellerPubKey> OP_SWAP
  OP_IF
    2 OP_SWAP <buyerPubKey> 2 OP_CHECKMULTISIG
  OP_ELSE
    <claimDelay> OP_CHECKSEQUENCEVERIFY OP_DROP OP_CHECKSIG
  OP_ENDIF