MakerProcessesInputsForDepositTxRequest: Merge all error catching cases
We have `failed(...)` methods for `String`s and `Throwable`s. The
`failed(Throwable t)` extracts the `Throwable`'s message and appends it
to the `errorMessage`. Therefore, both catch clauses are identical. We
can remove the other catch clause because the most outer catch clause is
identical to the inner one.

Relates to #7423