Fix quadratic time bug when syncing BM accounting data
Avoid doing linear scans of the entire list of accounting blocks every
time a new block is added (once to check for duplicates and once to get
the block with max height in 'BurningManAccountingStore::getLastBlock',
which JProfiler reveals to be the main bottleneck during syncs). Do this
by enforcing the invariant that the block heights are contiguous and
start from 'BurningManAccountingService.EARLIEST_BLOCK_HEIGHT', so that
there is a 1-to-1 mapping between heights and list indices, permitting
direct lookup by index. Also use a backing ArrayList instead of a linked
list, since the only disadvantage of the former is the failure to free
surplus capacity when the list is cleared (but that amounts to a few MB
at most, and it is otherwise more memory efficient).

(The above invariant needs to be additionally enforced in the store
constructor, so discard the first and all subsequent non-connecting
blocks and log an error. That was the only place in the code where it
could have been broken, as a result of corrupted store persistence.)